requirements.txt
pdf2docx>=0.5.8
python-docx>=1.1.0
PyMuPDF>=1.24.0
Pillow>=10.0.0
